+++ nuitka/MainControl.py
@@ -256,9 +256,7 @@ def makeSourceDirectory( main_module ):
 
     writeSourceCode(
         filename    = Utils.joinpath( source_dir, "__helpers.hpp" ),
-        source_code = CodeGeneration.generateHelpersCode(
-            context = global_context
-        )
+        source_code = CodeGeneration.generateHelpersCode()
     )
 
     module_hpp_include = [